lambda-functions-gcp-scale-image/requirements.txt
google-cloud-vision==1.0.0
google-cloud-storage==1.28.1
Wand==0.6.1
google-cloud-datastore==1.12.0
google-cloud==0.34.0